requirements.txt
Django==1.5.4
Markdown==2.3.1
Pygments==1.6
South==0.8.2
beautifulsoup4==4.3.2
dj-database-url==0.2.2
dj-static==0.0.5
django-filter==0.7
django-picklefield==0.3.0
django-toolbelt==0.0.1
djangorestframework==2.3.8
gunicorn==18.0
htmllaundry==2.0
jsonfield==0.9.19
lxml==3.2.3
psycopg2==2.5.1
python-dateutil==2.1
six==1.4.1
static==0.4
wsgiref==0.1.2
